// frontend/sdk/models/ScoreboardPlayerTotalDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sopranino API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ScoreboardPlayerTotalDto
 */
export interface ScoreboardPlayerTotalDto {
    /**
     * 
     * @type {string}
     * @memberof ScoreboardPlayerTotalDto
     */
    userId: string;
    /**
     * 
     * @type {string}
     * @memberof ScoreboardPlayerTotalDto
     */
    displayName: string;
    /**
     * 
     * @type {string}
     * @memberof ScoreboardPlayerTotalDto
     */
    avatarUrl?: string;
    /**
     * Total score across all rounds
     * @type {number}
     * @memberof ScoreboardPlayerTotalDto
     */
    totalScore: number;
}

/**
 * Check if a given object implements the ScoreboardPlayerTotalDto interface.
 */
export function instanceOfScoreboardPlayerTotalDto(value: object): value is ScoreboardPlayerTotalDto {
    if (!('userId' in value) || value['userId'] === undefined) return false;
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    if (!('totalScore' in value) || value['totalScore'] === undefined) return false;
    return true;
}

export function ScoreboardPlayerTotalDtoFromJSON(json: any): ScoreboardPlayerTotalDto {
    return ScoreboardPlayerTotalDtoFromJSONTyped(json, false);
}

export function ScoreboardPlayerTotalDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ScoreboardPlayerTotalDto {
    if (json == null) {
        return json;
    }
    return {
        
        'userId': json['userId'],
        'displayName': json['displayName'],
        'avatarUrl': json['avatarUrl'] == null ? undefined : json['avatarUrl'],
        'totalScore': json['totalScore'],
    };
}

export function ScoreboardPlayerTotalDtoToJSON(json: any): ScoreboardPlayerTotalDto {
    return ScoreboardPlayerTotalDtoToJSONTyped(json, false);
}

export function ScoreboardPlayerTotalDtoToJSONTyped(value?: ScoreboardPlayerTotalDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'userId': value['userId'],
        'displayName': value['displayName'],
        'avatarUrl': value['avatarUrl'],
        'totalScore': value['totalScore'],
    };
}